Stop exposing user classes directly under locust_plugins.users, and move listeners into their own submodule. This is done so that not all modules have to be read, giving us faster start up times and less irrelevant code being loaded, eliminating the need to install ALL dependencies if someone just needs to use a part of locust-plugins.
You will need to update your code from for example:
from locust_plugins.users import PlaywrightUser
to
from locust_plugins.users.playwright import PlaywrightUser

Also move listeners into their own submodule (e.g. locust_plugins.listeners.appinsights instead of locust_plugins.appinsights). Remove the Print listener (it is now available in locust.debug.PrintListener)

Selenium has been removed as a dependency, you'll need to install it manually from now on (because it conflicts with Playwright)
