+++ locust_plugins/listeners/appinsights.py
